TE-238 Restructured and added proper examples
