+++ src/components/collections/Grid/GridColumn/Readme.md
@@ -2,6 +2,10 @@
 <Grid>
  <GridRow>
    <GridColumn>
+    First column
+   </GridColumn>
+   <GridColumn>
+    Second column
    </GridColumn>
  </GridRow>
 </Grid>
