+++ src/components/collections/Grid/Readme.md
@@ -2,6 +2,18 @@
 <Grid>
  <GridRow>
    <GridColumn>
+    First Row - First Column
+   </GridColumn>
+   <GridColumn>
+    First Row - Second Column
+   </GridColumn>
+ </GridRow>
+ <GridRow>
+   <GridColumn>
+    Second Row - First Column
+   </GridColumn>
+   <GridColumn>
+    Second Row - Second Column
    </GridColumn>
  </GridRow>
 </Grid>
