+++ src/components/collections/Grid/constants.js
@@ -1,0 +1,2 @@
+export const COLUMNS = 12;
+export const FULL_WIDTH_COLUMN = 16;
