+++ src/components/collections/Grid/index.js
@@ -1,1 +1,3 @@
 export { Component as Grid } from './component';
+
+export { COLUMNS, FULL_WIDTH_COLUMN } from './constants';
