--- src/components/collections/GridRow/Readme.md
@@ -1,8 +1,0 @@
-```jsx
-<Grid>
- <GridRow>
-   <GridColumn>
-   </GridColumn>
- </GridRow>
-</Grid>
-```
