+++ src/components/widgets/Container/component.spec.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { GridColumn } from 'collections/GridColumn';
+import { GridColumn } from 'collections/Grid/GridColumn';
+import { FULL_WIDTH_COLUMN } from 'collections/Grid';
 import { OwnerLogin } from 'widgets/OwnerLogin';
 
 import { Component as Container } from './component';
@@ -28,7 +29,7 @@ describe('<Container />', () => {
       const actual = column.props();
       expect(actual).toEqual(
         expect.objectContaining({
-          width: 16,
+          width: FULL_WIDTH_COLUMN,
         })
       );
     });
